backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
google-generativeai
python-dotenv==1.0.0
